requirements.txt
aiobotocore==2.1.0
aiohttp==3.8.1
aioitertools==0.8.0
aiosignal==1.2.0
antlr4-python3-runtime==4.8
anyio==3.4.0
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
awscli-local==0.17
bcrypt==3.2.0
boto3==1.20.30
botocore==1.23.24
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
commonmark==0.9.1
cryptography==36.0.1
deepdiff==5.7.0
dill==0.3.2
dnslib==0.9.16
dnspython==2.1.0
docker==5.0.0
dulwich==0.20.28
ecdsa==0.17.0
fastapi==0.70.1
fastapi-login==1.7.3
frozenlist==1.2.0
h11==0.12.0
httptools==0.3.0
idna==2.10
importlib-resources==5.4.0
jmespath==0.10.0
localstack==0.13.2.1
localstack-client==1.29
localstack-ext==0.13.2.7
localstack-plugin-loader==1.1.1
loguru==0.5.3
mangum==0.12.3
multidict==5.2.0
omegaconf==2.1.1
ordered-set==4.0.2
passlib==1.7.4
pbr==5.8.0
psutil==5.9.0
pyaes==1.6.1
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
Pygments==2.11.2
PyJWT==2.3.0
pynamodb==5.2.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.25.1
rich==10.16.2
rsa==4.8
s3transfer==0.5.0
semver==2.13.0
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
stevedore==3.5.0
tabulate==0.8.9
tailer==0.4.1
typing-extensions==4.0.1
urllib3==1.26.7
uvicorn==0.16.0
uvloop==0.16.0
watchgod==0.7
websocket-client==1.2.3
websockets==10.1
wrapt==1.13.3
yarl==1.7.2
zipp==3.7.0
